code cleanup stuff
